Feat : posts PostDetail PUT, DELETE 인가 추가

+++ posts/views.py
@@ -18,9 +18,23 @@ from rest_framework.response import Response
 from rest_framework import status
 from django.http import Http404
 
+# 페이지네이션
+from rest_framework.pagination import PageNumberPagination
+
+
+# 인가
+from config.permissions import IsWriterOrReadOnly
+
+
+class CustomPageNumberPagination(PageNumberPagination):
+    page_size = 5
+    page_size_query_param = "page_size"
+    max_page_size = 100
+
 
 # PostList(전체)
 class PostList(APIView):
+    # Tag로 필터링 + 페이지네이션 + 정렬 -> 영어 한국어 게시물 모두 response
     def get(self, request):
         interest = request.query_params.get("interest")
         order = request.query_params.get("order")
@@ -28,10 +42,16 @@ class PostList(APIView):
 
         # 영어 posts
         posts_en = Post.objects.all()
+        # 영어 정렬
+        if order:
+            posts_en = posts_en.order_by(order)
         serializer_en = PostSerializer(posts_en, many=True)
 
         # 한국어 posts
         posts_kr = Post_KR.objects.all()
+        # 한국어 정렬
+        if order:
+            posts_kr = posts_kr.order_by(order)
         serializer_kr = Post_KRSerializer(posts_kr, many=True)
 
         if interest:
@@ -45,21 +65,29 @@ class PostList(APIView):
 
             # 영어 posts 중 파싱된 interest 값을 포함하는 게시물 필터링
             posts_en = posts_en.filter(q)
-            serializer_en = PostSerializer(posts_en, many=True)
 
-            print(serializer_en.data)
+            # 정렬
+            if order:
+                posts_en = posts_en.order_by(order)
+
+            # 페이지네이션
+            paginator = CustomPageNumberPagination()
+            paginated_posts_en = paginator.paginate_queryset(posts_en, request)
+            serializer_en = PostSerializer(paginated_posts_en, many=True)
+
             # 한국어 posts 중 필터링된 영어 post와 짝꿍 게시물 필터링
             if serializer_en.data != []:
                 post_id_list = [item["post_id"] for item in serializer_en.data]
                 posts_kr = posts_kr.filter(post__post_id__in=post_id_list)
                 serializer_kr = Post_KRSerializer(posts_kr, many=True)
 
+        print(serializer_en.data)
         data = {"post_en": serializer_en.data, "post_kr": serializer_kr.data}
         return Response(data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        language = request.data.get("language")  # 작성한 언어 판단
         user_id = request.data.get("user_id")
-        language = request.data.get("language")
         title = request.data.get("title")
         content = request.data.get("content")
         interest = request.data.get("interest")
@@ -85,7 +113,7 @@ class PostList(APIView):
                     serializer_en.errors, status=status.HTTP_400_BAD_REQUEST
                 )
 
-            # translation을 파싱하여 한국어 post에 값 저장
+            # translation을 파싱하여(\n을 기준으로) 한국어 post에 값 저장
             if "\n" in translation:
                 translation_list = translation.split("\n", 1)
                 kr_title = translation_list[0].strip()
@@ -105,13 +133,14 @@ class PostList(APIView):
                     serializer_kr.errors, status=status.HTTP_400_BAD_REQUEST
                 )
 
+            # en, kr 한 번에 담아서 response
             data = {
                 "serializer_en": serializer_en.data,
                 "serializer_kr": serializer_kr.data,
             }
             return Response(data, status.HTTP_201_CREATED)
 
-        # 가져온 데이터의 language가 한국어일 경우 (translation == english)
+        # 가져온 데이터의 language가 한국어일 경우 (translation == english) language가 영어일 경우와 반대
         else:
             if "\n" in translation:
                 translation_list = translation.split("\n", 1)
@@ -157,6 +186,8 @@ class PostList(APIView):
 
 # PostDetail(특정 값)
 class PostDetail(APIView):
+    permission_classes = [IsWriterOrReadOnly]
+
     def get(self, request, id):
         # 게시물 불러오기
         post_en = get_object_or_404(Post, post_id=id)
@@ -176,6 +207,8 @@ class PostDetail(APIView):
 
     def put(self, request, id):
         post = get_object_or_404(Post, post_id=id)
+        request.data["user_id"] = request.user.id
+        self.check_object_permissions(request, post)  # 작성자가 같은지 체크
         serializer = PostSerializer(post, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -185,5 +218,7 @@ class PostDetail(APIView):
 
     def delete(self, request, id):
         post = get_object_or_404(Post, post_id=id)
+        request.data["user_id"] = request.user.id
+        self.check_object_permissions(request, post)  # 작성자가 같은지 체크
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
